+++ CustomTooltips_LibActionButton/CustomTooltips_LibActionButton.lua
@@ -7,7 +7,7 @@ if not LibActionButton then
 	return
 end
 
-local function Button_OnEnter_Hook(self)	
+local function SetTooltip_Hook(self)	
 	local type, action = self:GetAction()
 	
 	local macroIndex
@@ -28,6 +28,22 @@ local function Button_OnEnter_Hook(self)
 	end
 end
 
+-- Hook the :SetTooltip method of the button's metatable index
+local function HookMeta(button) 
+	local methods = getmetatable(button).__index
+	
+	hooksecurefunc(methods, "SetTooltip", SetTooltip_Hook)
+end
+
 LibActionButton.RegisterCallback("CustomTooltips_LibActionButton", "OnButtonCreated", function(_, button)
-	button:HookScript("OnEnter", Button_OnEnter_Hook)
+	-- Only hook the :SetTooltip methods once
+	LibActionButton.UnregisterCallback("CustomTooltips_LibActionButton", "OnButtonCreated")
+
+	button:SetState(0, "action", 1) -- Set the button's kind to "action" and its metatable to Action
+	HookMeta(button) -- Hook the Action metatable
+	
+	button:SetState(0, "macro", 1) -- Set the button's kind to "macro" and its metatable to Macro
+	HookMeta(button) -- Hook the Macro metatable
+	
+	button:ClearStates() -- Clear the button's states
 end)